requirements.txt
# Core dependencies for Terragon SDLC Framework
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
psutil>=5.8.0
requests>=2.25.0

# Scientific computing and ML
scipy>=1.7.0
scikit-learn>=1.0.0

# Async support
aiofiles>=0.7.0
httpx>=0.24.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Code quality
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0
mypy>=1.0.0
pylint>=2.15.0

# Security scanning
bandit>=1.7.0
safety>=2.0.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.18.0

# Development tools
pre-commit>=2.20.0
tox>=4.0.0

# Optional ML dependencies
torch>=1.12.0; extra == "ml"
tensorflow>=2.9.0; extra == "ml"

# Optional quantum dependencies  
qiskit>=0.39.0; extra == "quantum"
cirq>=0.15.0; extra == "quantum"

# Container and cloud
docker>=6.0.0; extra == "docker"
kubernetes>=24.0.0; extra == "k8s"